# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/clients/products": {
            "get": {
                "description": "Get all products by all category frontned from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Get products by all category for frontned",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ProductCategory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    }
                }
            }
        },
        "/v1/clients/products/:slug": {
            "get": {
                "description": "Get product by specific category for frontend from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Get product by specific category for frontend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug category",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/products": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get product data from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ProductBrand"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new product product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create product",
                "parameters": [
                    {
                        "description": "Payload format",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Product"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "deleted": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    }
                }
            }
        },
        "/v1/products/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get product product by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Product"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "deleted": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update product data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload format",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Product"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "deleted": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete product data from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete product data by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BasicProductResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ClientProduct": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string",
                    "example": "xxxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "description": {
                    "type": "string",
                    "example": "Some description"
                },
                "group": {
                    "type": "string",
                    "example": "Telkomsel"
                },
                "id": {
                    "type": "string",
                    "example": "xxxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "is_discount": {
                    "type": "boolean",
                    "example": false
                },
                "label": {
                    "type": "string",
                    "example": "Telkomsel 5GB Super 24H"
                },
                "price": {
                    "type": "integer",
                    "example": 50000
                },
                "stock": {
                    "type": "integer",
                    "example": 30
                }
            }
        },
...
        "dto.ProductCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "xxxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "label": {
                    "type": "string",
                    "example": "Some Category"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ClientProduct"
                    }
                },
                "slug": {
                    "type": "string",
                    "example": "some-category"
                }
            }
        },
        "models.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "billed": {
                    "type": "integer"
                },
                "charged": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "methodDetails": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.ProductBrandCategory"
                },
                "category_id": {
                    "type": "string"
                },
                "deleted": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "some description"
                },
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxx"
                },
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "is_discount": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "Some Product"
                },
                "price": {
                    "type": "integer",
                    "example": 30000
                },
                "stock": {
                    "type": "integer",
                    "example": 30
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ProductBrand": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "some description"
                },
                "icon_path": {
                    "type": "string",
                    "example": "/path/to/file"
                },
                "id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "isAvailable": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "some name"
                },
                "product_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductCategory"
                    }
                }
            }
        },
        "models.ProductBrandCategory": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "category_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "models.ProductCategory": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "some description"
                },
                "icon": {
                    "type": "string",
                    "example": "mdi-some-category"
                },
                "id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Some Name"
                },
                "slug": {
                    "type": "string",
                    "example": "some-name"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "some description"
                },
                "id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "payment": {
                    "$ref": "#/definitions/models.Payment"
                },
                "product_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "updated": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                }
            }
        },
...
        "request.CreateProductRequest": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "category_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "description": {
                    "type": "string",
                    "example": "some description"
                },
                "name": {
                    "type": "string",
                    "example": "Some Product"
                },
                "price": {
                    "type": "integer",
                    "example": 30000
                },
                "stock": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
        "request.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "category_id": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "description": {
                    "type": "string",
                    "example": "some description"
                },
                "is_discount": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Some Product"
                },
                "price": {
                    "type": "integer",
                    "example": 30000
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "stock": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
        "response.BasicProductResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                },
                "status": {
                    "type": "string",
                    "example": "fail"
                }
            }
        },
        "response.BasicProductSuccessResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}